Fixed compile errors under MacOS, still need to find a fix for "Extra qualification on member" thing through...

+++ include/Framework/Logging/Logger.hpp
@@ -49,7 +49,7 @@ namespace bpf
         {
             String res = "";
 
-            for (size_t i = 0; i < format.Size(); ++i)
+            for (uint32 i = 0; i < format.Size(); ++i)
             {
                 if (format[i] == '@' && ((i + 1 == format.Size()) || (i + 1 != format.Size() && format[i + 1] != '@')))
                     return (res + String::ValueOf(std::forward<Arg>(arg))
